requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
Flask==3.0.1
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pdf2image==1.17.0
pillow==10.2.0
pycparser==2.21
Werkzeug==3.0.1
